fix(post-959): close P0–P3 audit residuals in one PR
Security:
- Durable webhook rate keys use trusted-proxy client resolution (no XFF spoof)
- JOE-952 covers /storage/export + session messages; localAdmin dual-intent uniform
- Rate-limiter twin parity test (shared vs gateway-channel)

Maintainability:
- Extract work-presentation + config-schema/normalize under soft 1500 LOC
- Wiki jobs-subject split; widen import-cycle SCAN_ROOTS to packages/shared
- Progressive createLocalWorkspaceSessionPort factory

Ops/docs/P0 package:
- Private-beta package inventory + no-go fail-closed validator
- Durable V2 doc drift, historical audit banner, HA/channel freeze dispositions
- DOMPurify ≥3.4.12 override + notices refresh
