+++ pnpm-workspace.yaml
@@ -38,7 +38,7 @@ onlyBuiltDependencies:
 
 overrides:
   "@ungap/structured-clone": 1.3.1
-  "dompurify@<3.4.11": ">=3.4.11"
+  "dompurify@<3.4.12": ">=3.4.12"
   electron-builder-squirrel-windows: 26.8.1
   "form-data@<4.0.6": ">=4.0.6"
   "hono@<4.12.25": ">=4.12.25"
