+++ products/gateway/src/__tests__/module-boundaries.test.ts
@@ -49,9 +49,9 @@ describe('module boundary budget', () => {
           edgeDelta: 12,
         },
         runtimeRisk: {
-          // Types extraction for work-store/environments LOC budgets (audit 2026-07-21).
-          moduleDelta: 12,
-          edgeDelta: 80,
+          // Types extraction (audit 2026-07-21) + work/config façade LOC split (post-959).
+          moduleDelta: 15,
+          edgeDelta: 110,
         },
       },
     })
