+++ products/gateway/src/__tests__/webhook-rate-limit.test.ts
@@ -33,11 +33,21 @@ describe('Durable inbound webhook rate limit', () => {
     expect(blocked.ok).toBe(false)
   })
 
-  it('builds stable rate keys from remote address', () => {
+  it('builds stable rate keys from socket remote address', () => {
     expect(inboundWebhookRateKey('whatsapp', { socket: { remoteAddress: '203.0.113.10' } })).toBe('whatsapp:203.0.113.10')
+  })
+
+  it('ignores spoofed X-Forwarded-For unless the socket peer is a trusted proxy', () => {
+    // Untrusted peer: client-supplied XFF must not mint a new bucket identity.
     expect(inboundWebhookRateKey('discord', {
-      socket: { remoteAddress: '127.0.0.1' },
-      headers: { 'x-forwarded-for': '198.51.100.2, 127.0.0.1' },
-    })).toBe('discord:198.51.100.2')
+      socket: { remoteAddress: '203.0.113.50' },
+      headers: { 'x-forwarded-for': '198.51.100.2, 203.0.113.50' },
+    }, { trustedProxyCidrs: [] })).toBe('discord:203.0.113.50')
+
+    // Trusted proxy peer: walk the chain to the client hop.
+    expect(inboundWebhookRateKey('discord', {
+      socket: { remoteAddress: '10.0.0.1' },
+      headers: { 'x-forwarded-for': '198.51.100.2, 10.0.0.1' },
+    }, { trustedProxyCidrs: ['10.0.0.0/8'] })).toBe('discord:198.51.100.2')
   })
 })
