+++ products/wiki/packages/postgres-runtime/src/jobs.ts
@@ -8,6 +8,7 @@ import { migratePostgresRuntime } from "./migrations.ts";
 import { jobStatusCount, runStatusCount } from "./queries.ts";
 import { dateStringField, jsonb, optionalSanitizedRunInput, runFromRow, stringField } from "./rows.ts";
 import { DEFAULT_STALE_RUN_MAX_RUNTIME_MS, RUNTIME_SOURCE_COMMIT, type CountRow, type JobAttemptRow, type PostgresQuery, type PostgresRunCancellationOptions, type PostgresRunCancellationResult, type PostgresRunJobInput, type PostgresRunQueueAdapter, type PostgresRunReaperOptions, type PostgresRunReaperResult, type PostgresRuntimeOptions, type PostgresRuntimeQueueHealth, type PostgresSql, type RunRow } from "./types.ts";
+import { optionalRunSubjectPaths, optionalEventSubjectPaths, optionalRunSensitivity, optionalEventSensitivity, runSubjectPaths, runSensitivity } from "./jobs-subject.ts";
 
 export async function createPostgresRunQueue(root: string, options: PostgresRuntimeOptions = {}): Promise<PostgresRunQueueAdapter> {
   const databaseUrl = resolvePostgresDatabaseUrl(options);
@@ -553,40 +554,6 @@ class PostgresRunQueue implements PostgresRunQueueAdapter {
   }
 }
 
-function optionalRunSubjectPaths(runType: string, explicitPaths: string[] | undefined): { subject_paths: string[] } | {} {
-  const subjectPaths = runSubjectPaths(runType, explicitPaths);
-  return subjectPaths === undefined ? {} : { subject_paths: subjectPaths };
-}
-
-function optionalEventSubjectPaths(runType: string, explicitPaths: string[] | undefined): { subject_paths: string[] } | {} {
-  const subjectPaths = runSubjectPaths(runType, explicitPaths);
-  return subjectPaths === undefined ? {} : { subject_paths: subjectPaths };
-}
-
-function optionalRunSensitivity(runType: string): { sensitivity: OpenWikiSectionVisibility } | {} {
-  const sensitivity = runSensitivity(runType);
-  return sensitivity === undefined ? {} : { sensitivity };
-}
-
-function optionalEventSensitivity(runType: string): { sensitivity: OpenWikiSectionVisibility } | {} {
-  const sensitivity = runSensitivity(runType);
-  return sensitivity === undefined ? {} : { sensitivity };
-}
-
-function runSubjectPaths(runType: string, explicitPaths: string[] | undefined): string[] | undefined {
-  const paths = [...(explicitPaths ?? []), ...(runType === "source.fetch" ? ["sources/manifests", "sources/raw"] : [])]
-    .map((entry) => entry.trim())
-    .filter((entry, index, values) => entry.length > 0 && values.indexOf(entry) === index);
-  return paths.length === 0 ? undefined : paths;
-}
-
-function runSensitivity(runType: string): OpenWikiSectionVisibility | undefined {
-  return runType === "source.fetch" ? "internal" : undefined;
-}
-
-// postgres.js types the in-transaction handle (passed to `sql.begin`) as a distinct generic
-// instantiation that does not structurally unify with PostgresSql, even though both accept the
-// same tagged-template queries. Narrow it once, here, instead of at every transaction call site.
 function transactionQuery(tx: unknown): PostgresQuery {
   return tx as PostgresQuery;
 }
